complete database population, export and testing with db switch

+++ scripts/start-substrate.sh
@@ -44,22 +44,24 @@ fi
 SUBSTRATE_PID=$(docker exec $SUBSTRATE_CONTAINER_NAME sh -c "ps aux | grep -v grep | grep substrate-contracts-node | awk '{print \$2}'")
 MAIN_PID=$(docker exec $SUBSTRATE_CONTAINER_NAME sh -c "ps -ef | grep \"0\ssleep infinity\" | awk 'FNR == 1 {print \$2}'")
 # restart chain with no data
-if [[ $RESTART_CHAIN == true ]]; then
+if [[ $RESTART_CHAIN == true || $TEST_DB == true ]]; then
   if [[ -n "$SUBSTRATE_PID" ]]; then
     for pid in $SUBSTRATE_PID; do
       echo "Killing existing substrate pid $pid..."
       docker exec "$SUBSTRATE_CONTAINER_NAME" kill -9 "$pid"
     done
   fi
-  docker exec -d "$SUBSTRATE_CONTAINER_NAME" rm -rf ./chain-data
+  if [[ $RESTART_CHAIN == true ]]; then
+    docker exec -d "$SUBSTRATE_CONTAINER_NAME" rm -rf ./chain-data
+  fi
 fi
 
 # switch db
 if [[ $TEST_DB == true ]]; then
-  if [[ -n "$SUBSTRATE_PID" ]]; then
-    echo "Killing existing substrate..."
-    docker exec "$SUBSTRATE_CONTAINER_NAME" kill -9 "$SUBSTRATE_PID"
-  fi
+  # if [[ -n "$SUBSTRATE_PID" ]]; then
+  #   echo "Killing existing substrate..."
+  #   docker exec "$SUBSTRATE_CONTAINER_NAME" kill -9 "${SUBSTRATE_PID//$'\n'/ }"
+  # fi
   docker exec -d "$SUBSTRATE_CONTAINER_NAME" rm -rf ./chain-test
   docker cp ./.chain-test/. "$SUBSTRATE_CONTAINER_NAME":/chain-test
   docker exec -d "$SUBSTRATE_CONTAINER_NAME" bash -c "substrate-contracts-node --dev -d ./chain-test --unsafe-ws-external --rpc-external --prometheus-external -lerror,runtime::contracts=debug >> /proc/$MAIN_PID/fd/1  2>&1"
